+++ trunk/Server/server.h
@@ -20,7 +20,7 @@ public:
 
 	static bool CALLBACK OnConnection(ConnID connId);
 	static void CALLBACK OnDisconnect(ConnID connId);
-	static void CALLBACK OnData(ConnID connId, uint16 iLen, char* pBuf);
+	static void CALLBACK OnData(ConnID connId, uint32 iLen, char* pBuf);
 	static void CALLBACK OnConnectFailed(void*);
 
 	static std::string GetNickName(Connection* pConnection);
